+++ demo/EmployeeField/index.js
@@ -10,7 +10,9 @@ import 'salt-context';
 import './EmployeeFieldDemo.styl';
 import Demo from './EmployeeFieldDemo';
 
-window.FastClick && FastClick.attach(document.body);
+if (window.FastClick) {
+  window.FastClick.attach(document.body);
+}
 
 // 渲染demo
 export default Demo;